synergyway_test_task/requirements.txt
amqp==5.1.1
asgiref==3.7.2
celery==5.2.6
django-extensions==3.2.3
django-seed==0.3.1
django-unfold==0.42.0
Django==4.2.6
email-validator==2.2.0
factory-boy==3.3.0
fastapi-pagination==0.12.32
fastapi-restful[all]==0.6.0
fastapi==0.115.6
httpx==0.28.1
google-api-python-client==2.114.0
google-auth-oauthlib==1.2.0
Pillow==9.1.0
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
pydantic[email]==2.10.3
pytest-django==4.8.0
pytest-asyncio==0.25.2
pytest-factoryboy==2.7.0
pytest==8.3.2
python-dotenv==0.21.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
redis==5.0.1
typing-inspect==0.9.0
uvicorn==0.23.2
fastapi_utils==0.8.0
ruff==0.11.3